Removing config.json file from tree.
This file contains sensitive information about credentials that may be
in production.
From now one, anybody that wants to setup the project on their machine
and/or deploy to production must make sure that the config/config.json
file is present and matches with the config/config.json.example one with
the proper credentials set.